backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
supabase==2.7.4
httpx==0.27.0
reportlab==4.0.7
Werkzeug==3.0.1
